test for both DurableIteratorObject & DurableIteratorObjectHandler
